Improve JS handling of non-ICAO addresses
Add more special case handling for addresses that start with '~'. We
aren't able to look these up on Airframes or FlightAware, so remove
those links. Additionally, we shouldn't try to load a JSON DB for the
'~' prefix, so add a special case check for that.

Fix a variety of places near the code I was editing that was using tabs
instead of spaces for indentation for consistency.

Improve the Node.js module setup as it made it easier to test the
changes I made here without a browser.